# backend/docs/swagger.yaml
basePath: /api
definitions:
  domain.AdminStats:
    properties:
      activeCoursesCount:
        type: integer
      activeUsers:
        type: integer
      newUsersLastMonth:
        type: integer
      newUsersLastWeek:
        type: integer
      storageUsed:
        description: 单位：字节
        type: integer
      systemStartTime:
        type: string
      totalCourses:
        type: integer
      totalCustomTracks:
        type: integer
      totalTracks:
        type: integer
      totalUsers:
        type: integer
    type: object
  domain.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      error:
        type: string
    type: object
...
  main.AudioMetadata:
    properties:
      artist:
        type: string
      bitRate:
        type: integer
      bookmarks:
        items:
          $ref: '#/definitions/main.AudioBookmark'
        type: array
      channels:
        type: integer
      customData:
        additionalProperties: true
        type: object
      duration:
        type: number
      format:
        type: string
      sampleRate:
        type: integer
      spectralData:
        items:
          type: number
        type: array
      title:
        type: string
      trackId:
        type: string
      transcriptData:
        $ref: '#/definitions/main.AudioTranscript'
      waveformData:
        items:
          type: number
        type: array
    type: object
  main.AudioTokenRequest:
    properties:
      action:
        enum:
        - stream_audio
        - get_metadata
        type: string
      courseId:
        type: string
      trackId:
        type: string
      unitId:
        type: string
    required:
    - action
    - courseId
    - trackId
    - unitId
    type: object
  main.AudioTokenResponse:
    description: 音频访问令牌响应
    properties:
      expiresAt:
        description: 过期时间戳
        example: 1609459200
        type: integer
      token:
        description: 访问令牌
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  main.ErrorResponse:
    description: API错误响应
    properties:
      error:
        description: 错误信息
        example: 未授权访问
        type: string
    type: object
...
  main.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 提供语言学习音频播放器的后端API服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 语言学习音频播放器API
  version: "1.0"
paths:
  /api/admin/stats:
    get:
      consumes:
      - application/json
      description: 获取系统运行的统计数据
      produces:
      - application/json
      responses:
        "200":
          description: 系统统计信息
          schema:
            $ref: '#/definitions/domain.AdminStats'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取系统统计信息
      tags:
      - admin
  /api/admin/users:
    get:
      consumes:
      - application/json
      description: 获取所有系统用户的列表，支持分页
      parameters:
      - description: 页码，默认为1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认为10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取系统用户列表
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: 创建新的系统用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建系统用户
      tags:
      - admin
  /api/admin/users/{userId}:
    delete:
      consumes:
      - application/json
      description: 删除指定的系统用户
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除系统用户
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: 更新现有系统用户的信息
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: string
      - description: 用户更新信息
        in: body
        name: user
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新系统用户
      tags:
      - admin
  /api/audio/adaptive/{trackId}:
    get:
      consumes:
      - application/json
      description: 基于网络状况和客户端能力自动选择合适质量的音频流
      parameters:
      - description: 音轨ID
        in: path
        name: trackId
        required: true
        type: string
      - description: 访问令牌
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 音频文件
          schema:
            type: file
        "400":
          description: 请求无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 音频文件不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 自适应音频流
      tags:
      - audio
  /api/audio/metadata/{trackId}:
    get:
      consumes:
      - application/json
      description: 获取音频文件的元数据信息，包括波形数据
      parameters:
      - description: 音轨ID
        in: path
        name: trackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 音频元数据
          schema:
            $ref: '#/definitions/main.AudioMetadata'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 音频文件不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取音频元数据
      tags:
      - audio
  /api/audio/stream/{trackId}:
    get:
      consumes:
      - application/json
      description: 流式传输音频内容，支持范围请求
      parameters:
      - description: 音轨ID
        in: path
        name: trackId
        required: true
        type: string
      - description: 访问令牌
        in: query
        name: token
        required: true
        type: string
      - description: 范围请求头
        in: header
        name: Range
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 音频文件
          schema:
            type: file
        "206":
          description: 部分内容(范围请求)
          schema:
            type: file
        "400":
          description: 请求无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 音频文件不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 获取音频流
      tags:
      - audio
  /api/audio/token/{trackId}:
    get:
      consumes:
      - application/json
      description: 生成用于访问音频资源的临时令牌
      parameters:
      - description: 音轨ID
        in: path
        name: trackId
        required: true
        type: string
      - description: 令牌请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.AudioTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 令牌响应
          schema:
            $ref: '#/definitions/main.AudioTokenResponse'
        "400":
          description: 请求无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 音频文件不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取音频访问令牌
      tags:
      - audio
  /api/audio/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传用户自定义音频文件
      parameters:
      - description: 音频文件
        in: formData
        name: file
        required: true
        type: file
      - description: 音频标题
        in: formData
        name: title
        required: true
        type: string
      - description: 课程ID
        in: formData
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: formData
        name: unitId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 上传成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "413":
          description: 文件过大
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 上传自定义音频
      tags:
      - audio
  /api/audio/user-tracks:
    get:
      consumes:
      - application/json
      description: 获取当前用户上传的所有音频文件
      produces:
      - application/json
      responses:
        "200":
          description: 音频列表
          schema:
            items:
              $ref: '#/definitions/main.TrackResponse'
            type: array
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户音频列表
      tags:
      - audio
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 验证用户凭据并返回JWT令牌
      parameters:
      - description: 登录信息
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/main.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/main.LoginResponse'
        "400":
          description: 请求无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 用户登录
      tags:
      - auth
...
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账户并返回JWT令牌
      parameters:
      - description: 注册信息
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/main.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            $ref: '#/definitions/main.LoginResponse'
        "400":
          description: 请求无效或邮箱已被注册
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 用户注册
      tags:
      - auth
  /api/auth/validate:
    get:
      consumes:
      - application/json
      description: 验证当前用户会话状态并返回用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 无效会话
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 验证用户会话
      tags:
      - auth
  /api/courses:
    get:
      consumes:
      - application/json
      description: 获取当前用户可访问的所有课程
      produces:
      - application/json
      responses:
        "200":
          description: 课程列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户课程列表
      tags:
      - courses
  /api/courses/{courseId}:
    get:
      consumes:
      - application/json
      description: 获取指定课程的详细信息
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 课程详情
          schema:
            $ref: '#/definitions/main.CourseInfo'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 课程不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取课程详情
      tags:
      - courses
  /api/courses/{courseId}/units:
    get:
      consumes:
      - application/json
      description: 获取指定课程的所有单元
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 课程单元列表
          schema:
            items:
              $ref: '#/definitions/main.CourseUnitInfo'
            type: array
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 课程不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取课程单元列表
      tags:
      - courses
  /api/courses/{courseId}/units/{unitId}/custom-tracks:
    get:
      consumes:
      - application/json
      description: 获取用户在指定课程和单元下的自定义音轨
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 自定义音轨列表
          schema:
            items:
              $ref: '#/definitions/main.CustomTrack'
            type: array
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 获取用户自定义音轨
      tags:
      - custom-tracks
    post:
      consumes:
      - multipart/form-data
      description: 上传并添加新的自定义音轨
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      - description: 音频文件
        in: formData
        name: audioFile
        required: true
        type: file
      - description: 音轨标题
        in: formData
        name: title
        required: true
        type: string
      - description: 中文标题
        in: formData
        name: chineseName
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 添加成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "413":
          description: 文件过大
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 添加自定义音轨
      tags:
      - custom-tracks
  /api/courses/{courseId}/units/{unitId}/custom-tracks/{trackId}:
    delete:
      consumes:
      - application/json
      description: 删除用户创建的自定义音轨
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      - description: 音轨ID
        in: path
        name: trackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 无权限操作
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 音轨不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 删除自定义音轨
      tags:
      - custom-tracks
    patch:
      consumes:
      - application/json
      description: 更新自定义音轨的标题和中文名称
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      - description: 音轨ID
        in: path
        name: trackId
        required: true
        type: string
      - description: 更新信息
        in: body
        name: track
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 无权限操作
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 音轨不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 更新自定义音轨
      tags:
      - custom-tracks
  /api/courses/{courseId}/units/{unitId}/custom-tracks/{trackId}/reorder:
    post:
      consumes:
      - application/json
      description: 调整自定义音轨的排序顺序
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      - description: 音轨ID
        in: path
        name: trackId
        required: true
        type: string
      - description: 排序信息
        in: body
        name: order
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 排序成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 无权限操作
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 音轨不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 重排序自定义音轨
      tags:
      - custom-tracks
  /api/courses/{courseId}/units/{unitId}/tracks:
    get:
      consumes:
      - application/json
      description: 获取指定课程单元的所有音轨
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 单元ID
        in: path
        name: unitId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 音轨列表
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: 课程或单元不存在
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取单元音轨列表
      tags:
      - courses
...
  /api/recent-tracks:
    get:
      consumes:
      - application/json
      description: 获取用户最近播放过的音轨列表
      produces:
      - application/json
      responses:
        "200":
          description: 最近播放列表
          schema:
            items:
              $ref: '#/definitions/main.RecentTrackInfo'
            type: array
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取最近播放记录
      tags:
      - tracks
  /api/track-progress:
    post:
      consumes:
      - application/json
      description: 保存用户的音轨播放位置和完成率
      parameters:
      - description: 进度信息
        in: body
        name: progress
        required: true
        schema:
          $ref: '#/definitions/main.TrackProgressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新音轨播放进度
      tags:
      - tracks
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"